Update of the backspace key for mouse click, so you can play on the computer and on your cell phone.
app. js: small change update of the backspace key for mouse click, so you can play on the computer and on your cell phone.
index.html: add the meta viewport user-scalable=no, so when the game is opened on the cell phone the zoom is disabled to work with the mouse click.